+++ sdk/ai/azure-ai-inference/tests/test_chat_completions_client.py
@@ -85,7 +85,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
                     max_tokens=321,
                     model="some-model-id",
                     presence_penalty=4.567,
-                    response_format=sdk.models.ChatCompletionsResponseFormatJSON(),
+                    response_format="json_object",
                     seed=654,
                     stop=["stop1", "stop2"],
                     stream=True,
@@ -122,7 +122,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
             max_tokens=321,
             model="some-model-id",
             presence_penalty=4.567,
-            response_format=sdk.models.ChatCompletionsResponseFormatJSON(),
+            response_format="json_object",
             seed=654,
             stop=["stop1", "stop2"],
             temperature=8.976,
@@ -197,7 +197,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
             max_tokens=768,
             model="some-other-model-id",
             presence_penalty=1.234,
-            response_format=sdk.models.ChatCompletionsResponseFormatText(),
+            response_format="text",
             seed=987,
             stop=["stop3", "stop5"],
             temperature=5.432,
@@ -256,7 +256,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
                     max_tokens=321,
                     model="some-model-id",
                     presence_penalty=4.567,
-                    response_format=sdk.models.ChatCompletionsResponseFormatJSON(),
+                    response_format="json_object",
                     seed=654,
                     stop=["stop1", "stop2"],
                     stream=True,
@@ -290,7 +290,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
         self._print_model_info_result(response1)
         self._validate_model_info_result(
             response1, "chat-completion"
-        )  # TODO: This should be ModelType.CHAT once the model is fixed
+        )  # TODO: This should be ModelType.CHAT_COMPLETION once the model is fixed
         client.close()
 
     @ServicePreparerChatCompletions()
@@ -306,7 +306,7 @@ class TestChatCompletionsClient(ModelClientTestBase):
         self._print_model_info_result(response1)
         self._validate_model_info_result(
             response1, "chat-completion"  # TODO: This should be chat_comletions according to REST API spec...
-        )  # TODO: This should be ModelType.CHAT once the model is fixed
+        )  # TODO: This should be ModelType.CHAT_COMPLETION once the model is fixed
 
         # Get the model info again. No network calls should be made here,
         # as the response is cached in the client.
@@ -492,7 +492,9 @@ class TestChatCompletionsClient(ModelClientTestBase):
             ],
         )
         self._print_chat_completions_result(response)
-        self._validate_chat_completions_result(response, ["juggling", "balls", "blue", "red", "green", "yellow"], True)
+        self._validate_chat_completions_result(
+            response, ["juggling", "balls", "blue", "red", "green", "yellow"], is_aoai=True
+        )
         client.close()
 
     # We use AOAI endpoint here because at the moment there is no MaaS model that supports
@@ -516,13 +518,15 @@ class TestChatCompletionsClient(ModelClientTestBase):
             ],
         )
         self._print_chat_completions_result(response)
-        self._validate_chat_completions_result(response, ["juggling", "balls", "blue", "red", "green", "yellow"], True)
+        self._validate_chat_completions_result(
+            response, ["juggling", "balls", "blue", "red", "green", "yellow"], is_aoai=True
+        )
         client.close()
 
     # We use AOAI endpoint here because at the moment MaaS does not support Entra ID auth.
     @ServicePreparerAOAIChatCompletions()
     @recorded_by_proxy
-    def test_chat_aoai_completions_with_entra_id_auth(self, **kwargs):
+    def test_aoai_chat_completions_with_entra_id_auth(self, **kwargs):
         client = self._create_aoai_chat_client(key_auth=False, **kwargs)
         messages = [
             sdk.models.SystemMessage(content="You are a helpful assistant answering questions regarding length units."),
@@ -530,7 +534,29 @@ class TestChatCompletionsClient(ModelClientTestBase):
         ]
         response = client.complete(messages=messages)
         self._print_chat_completions_result(response)
-        self._validate_chat_completions_result(response, ["5280", "5,280"], True)
+        self._validate_chat_completions_result(response, ["5280", "5,280"], is_aoai=True)
+        client.close()
+
+    @ServicePreparerAOAIChatCompletions()
+    @recorded_by_proxy
+    def test_aoai_chat_completions_with_structured_output(self, **kwargs):
+        client = self._create_aoai_chat_client(key_auth=True, **kwargs)
+        response_format = sdk.models.JsonSchemaFormat(
+            name="Test_JSON_Schema",
+            schema=ModelClientTestBase.OUTPUT_FORMAT_JSON_SCHEMA,
+            description="Describes a set of distances between locations",
+            strict=True,
+        )
+        print(type(response_format))
+        messages = [
+            sdk.models.SystemMessage(content="You are a helpful assistant answering questions on US geography"),
+            sdk.models.UserMessage(content="What's the distance between Seattle and Portland, as the crow flies?"),
+        ]
+        response = client.complete(messages=messages, response_format=response_format)
+        self._print_chat_completions_result(response)
+        self._validate_chat_completions_result(
+            response, ["distances", "location1", "Seattle", "location2", "Portland"], is_aoai=True, is_json=True
+        )
         client.close()
 
     # **********************************************************************************