+++ sdk/ai/azure-ai-inference/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Features Added
 
 * Added a client for Image Embeddings, named `ImageEmbeddingsClient`. See package README.md and new samples.
+* Added support for Chat Completions response message in JSON format that adheres to a given JSON schema. Also known
+as "structured output". See new samples `sample_chat_completions_with_structured_output.py` and
+`sample_chat_completions_with_structured_output_pydantic.py`.
+
+### Breaking Changes
+
+* If you previously configured your `ChatCompletionClient.complete()` call to output JSON format without a scheme, you have this in your code: `response_format=ChatCompletionsResponseFormatJSON()`. To maintain the same functionality, replace this with `response_format="json-object"`. We however recommend that you now switch to output JSON format with a provided schema if your AI model supports it.
 
 ### Bugs Fixed